+++ core/services/item_effects/add_wipe_bomb_attempts_effect.py
@@ -3,8 +3,8 @@ from typing import Dict, Any
 import json
 
 from .abstract_effect import AbstractItemEffect
-from ....core.domain.models import User, Item, UserBuff
-from ....core.utils import get_now
+from ...domain.models import User, Item, UserBuff
+from ...utils import get_now
 
 
 def get_end_of_day():
