+++ core/services/item_effects/refine_destruction_shield_effect.py
@@ -2,8 +2,8 @@ import json
 from typing import Dict, Any
 
 from .abstract_effect import AbstractItemEffect
-from ....core.domain.models import User, Item, UserBuff
-from ....core.utils import get_now
+from ...domain.models import User, Item, UserBuff
+from ...utils import get_now
 
 
 class RefineDestructionShieldEffect(AbstractItemEffect):
