+++ core/services/item_effects/reset_steal_cooldown_effect.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from typing import Dict, Any
 
 from .abstract_effect import AbstractItemEffect
-from ....core.domain.models import User, Item
+from ...domain.models import User, Item
 
 
 class ResetStealCooldownEffect(AbstractItemEffect):
